BORKED: start reworking ready for libril integration
This need to be reworked to be multithreaded and support synchronous QMI
messages to greatly simplify the code, there is also a lot of code
duplication which needs to be reduced.

This definitely doesn't compile as it's mid-rewrite.

A second refactor may follow to split the qrild_msg.c code into a
separate library so it can be more distinct from the rest of the
codebase and be integrated into future projects. It does a lot of heavy
lifting you might want in a QMI/QRTR implementation and should be easy
to implement with it's own data structure and simple threading in the
client.