big QMI rework and more impl
Start making use of the new struct-per-message for QMI messages, this
will make it wayyy easier to implement new QMI messages as we're getting
to the more complicated network oriented ones.

next up is getting and sending all the network and operator data so
android actually knows what's going on and hopefully finally detects the
SIM cards properly.